[AIRToAIE] Address review: narrowing cast, refeed predicate, test robustness
- cap_lock init: explicit static_cast<int> on the int64 max, consistent with
  the nearby static_cast<int>(wlockInit) and silencing -Wshorten-64-to-32.
- ping-pong skip: use getRefeedCount(buf) <= 1 instead of hasAttr(RefeedCount),
  matching the canonical "refeed active" idiom used elsewhere in the file and
  correctly handling the normalized (n<=1 / overflow) cases; reword comment to
  the established "count-free re-broadcast" terminology.
- test: match the refeed attribute loosely (CHECK-SAME) so it is robust to
  attribute reordering, and use CHECK-COUNT-1 + CHECK-NOT to actually enforce a
  single buffer instance on the memtile.

+++ mlir/lib/Conversion/AIRToAIESchedulingUtils.cpp
@@ -645,7 +645,8 @@ air::DMAAllocator::getOrCreateChainLockSet(AIE::BufferOp buf,
   // needs the cap_lock primed to N: the first writer acquires cap >= N, and the
   // single reader releases cap by 1 per re-send (N sends drain sig[last]=N and
   // restore cap=N). Default init=1 would deadlock the first writer's acq>=N.
-  int capInit = std::max<int64_t>(1, air::getRefeedCount(buf.getOperation()));
+  int capInit = static_cast<int>(
+      std::max<int64_t>(1, air::getRefeedCount(buf.getOperation())));
   cls.cap_lock = allocateLockOp(device, tile, /*init=*/capInit);
 
   // N init=0 signal locks for the writer→writer (or reader→reader)
